feat: restore org recovery coverage and stabilize E2E suites
