+++ .github/workflows/test-frontend.yaml
@@ -80,7 +80,7 @@ jobs:
             soloRequired: true
             backendRequired: true
             sharedEnvironment: true
-            workers: 4
+            workers: 6
     steps:
       - name: Harden Runner
         uses: step-security/harden-runner@f808768d1510423e83855289c910610ca9b43176 # v2.17.0
