+++ automation/tests/local-transactions/transactionDraftTests.test.ts
@@ -17,7 +17,7 @@ let loginPage: LoginPage;
 let transactionPage: TransactionPage;
 let isolationContext: ActivatedTestIsolationContext | null = null;
 
-test.describe('Transaction draft tests @local-transactions', () => {
+test.describe('Transaction draft account tests @local-transactions', () => {
   test.beforeAll(async () => {
     ({ app, window, isolationContext } = await setupLocalSuiteApp(test.info()));
   });
@@ -249,87 +249,6 @@ test.describe('Transaction draft tests @local-transactions', () => {
 
     const spenderAccountIdFromField = await transactionPage.getSpenderAccountId();
     expect(spenderAccountIdFromField).toContain(spenderAccountId);
-
-    await transactionPage.navigateToDrafts();
-    await transactionPage.deleteFirstDraft();
-  });
-
-  test('Verify draft transaction contains the saved info for create file tx', async () => {
-    await transactionPage.clickOnCreateNewTransactionButton();
-    await transactionPage.clickOnFileServiceLink();
-    await transactionPage.clickOnFileCreateTransaction();
-
-    const transactionMemoText = 'test memo';
-    const fileMemoText = 'file memo';
-    // const fileContent = 'test file content';
-
-    await transactionPage.fillInTransactionMemoUpdate(transactionMemoText);
-    await transactionPage.fillInFileMemo(fileMemoText);
-    // await transactionPage.fillInFileContent(fileContent);
-
-    await transactionPage.saveDraft();
-    await transactionPage.clickOnFirstDraftContinueButton();
-
-    const transactionMemoFromField = await transactionPage.getTransactionMemoText();
-    expect(transactionMemoFromField).toBe(transactionMemoText);
-
-    const fileMemoFromField = await transactionPage.getFileMemoTextFromField();
-    expect(fileMemoFromField).toBe(fileMemoText);
-
-    await transactionPage.navigateToDrafts();
-    await transactionPage.deleteFirstDraft();
-  });
-
-  test('Verify draft transaction contains the saved info for update file tx', async () => {
-    await transactionPage.ensureFileExists('test');
-    const fileId = await transactionPage.getFirsFileIdFromCache();
-    await transactionPage.clickOnCreateNewTransactionButton();
-    await transactionPage.clickOnFileServiceLink();
-    await transactionPage.clickOnUpdateFileSublink();
-
-    const transactionMemoText = 'test memo';
-    const fileMemoText = 'file memo';
-
-    await transactionPage.fillInTransactionMemoUpdate(transactionMemoText);
-    await transactionPage.fillInFileIdForUpdate(fileId?? '');
-    await transactionPage.fillInFileMemo(fileMemoText);
-
-    await transactionPage.saveDraft();
-    await transactionPage.clickOnFirstDraftContinueButton();
-
-    const transactionMemoFromField = await transactionPage.getTransactionMemoText();
-    expect(transactionMemoFromField).toBe(transactionMemoText);
-
-    const fileIdFromPage = await transactionPage.getFileIdFromUpdatePage();
-    expect(fileId).toBe(fileIdFromPage);
-
-    const fileMemoFromField = await transactionPage.getFileMemoTextFromField();
-    expect(fileMemoFromField).toBe(fileMemoText);
-
-    await transactionPage.navigateToDrafts();
-    await transactionPage.deleteFirstDraft();
-  });
-
-  test('Verify draft transaction contains the saved info for append file tx', async () => {
-    await transactionPage.ensureFileExists('test');
-    const fileId = await transactionPage.getFirsFileIdFromCache();
-    await transactionPage.clickOnCreateNewTransactionButton();
-    await transactionPage.clickOnFileServiceLink();
-    await transactionPage.clickOnAppendFileSublink();
-
-    const transactionMemoText = 'test memo';
-
-    await transactionPage.fillInTransactionMemoUpdate(transactionMemoText);
-    await transactionPage.fillInFileIdForAppend(fileId?? '');
-
-    await transactionPage.saveDraft();
-    await transactionPage.clickOnFirstDraftContinueButton();
-
-    const transactionMemoFromField = await transactionPage.getTransactionMemoText();
-    expect(transactionMemoFromField).toBe(transactionMemoText);
-
-    const fileIdFromPage = await transactionPage.getFileIdFromAppendPage();
-    expect(fileId).toBe(fileIdFromPage);
 
     await transactionPage.navigateToDrafts();
     await transactionPage.deleteFirstDraft();
