+++ automation/tests/local-transactions/transactionTests.test.ts
@@ -6,7 +6,6 @@ import {
   getOperatorKeyEnv,
   setupEnvironmentForTransactions,
 } from '../../utils/runtime/environment.js';
-import { Transaction } from '../../../front-end/src/shared/interfaces/index.js';
 import { createSeededLocalUserSession } from '../../utils/seeding/localUserSeeding.js';
 import {
   setupLocalSuiteApp,
@@ -20,7 +19,7 @@ let loginPage: LoginPage;
 let transactionPage: TransactionPage;
 let isolationContext: ActivatedTestIsolationContext | null = null;
 
-test.describe('Transaction execution tests @local-transactions', () => {
+test.describe('Transaction account execution tests @local-transactions', () => {
   test.beforeAll(async () => {
     ({ app, window, isolationContext } = await setupLocalSuiteApp(test.info()));
   });
@@ -241,178 +240,4 @@ test.describe('Transaction execution tests @local-transactions', () => {
     // const key = accountDetails.accounts[0]?.key?.key;
   });
 
-  test('Verify user can execute transfer tokens tx', async () => {
-    await transactionPage.ensureAccountExists();
-    const accountFromList = await transactionPage.getFirstAccountFromList();
-    const amountToBeTransferred = 1;
-    const transactionId = await transactionPage.transferAmountBetweenAccounts(
-      accountFromList,
-      amountToBeTransferred.toString(),
-    );
-
-    const transactionDetails: Transaction = await transactionPage.mirrorGetTransactionResponse(
-      transactionId ?? '',
-    );
-
-    const transactionType = transactionDetails?.name;
-    const allTransfer = transactionDetails?.transfers;
-    const amount = allTransfer.find(acc => acc.account === accountFromList)?.amount;
-    const result = transactionDetails?.result;
-    expect(transactionType).toBe('CRYPTOTRANSFER');
-    expect(amount).toBe(amountToBeTransferred * 100000000);
-    expect(result).toBe('SUCCESS');
-  });
-
-  test('Verify user can add the rest of remaining hbars to receiver accounts', async () => {
-    const amountToBeTransferred = 10;
-    const amountLeftForRestAccounts = 9;
-    await transactionPage.ensureAccountExists();
-    const receiverAccount = await transactionPage.getFirstAccountFromList();
-    await loginPage.waitForToastToDisappear();
-
-    await transactionPage.clickOnTransactionsMenuButton();
-    await transactionPage.clickOnCreateNewTransactionButton();
-    await transactionPage.clickOnTransferTokensTransaction();
-    await transactionPage.fillInTransferFromAccountId();
-    await transactionPage.fillInTransferAmountFromAccount(amountToBeTransferred.toString());
-    await transactionPage.fillInTransferToAccountId(receiverAccount);
-    await transactionPage.clickOnAddTransferFromButton();
-    await transactionPage.fillInTransferAmountToAccount(
-      (amountToBeTransferred - amountLeftForRestAccounts).toString(),
-    );
-    await transactionPage.clickOnAddTransferToButton();
-
-    await transactionPage.fillInTransferToAccountId(await transactionPage.getPayerAccountId());
-    await transactionPage.clickOnAddRestButton();
-
-    // Get HBAR amounts for the two accounts and verify rest is added up
-    const amounts = await transactionPage.getHbarAmountValueForTwoAccounts()
-    const firstReceiverAmount = amounts?.split(',')[0] ?? ''
-    const secondReceiverAmount = amounts?.split(',')[1] ?? ''
-    expect(firstReceiverAmount).toContain(
-      (amountToBeTransferred - amountLeftForRestAccounts).toString(),
-    );
-    expect(secondReceiverAmount).toContain(amountLeftForRestAccounts.toString());
-  });
-
-  test('Verify sign button is disabled when receiver amount is higher than payer amount when doing transfer tx', async () => {
-    await transactionPage.ensureAccountExists();
-    const receiverAccount = await transactionPage.getFirstAccountFromList();
-    await loginPage.waitForToastToDisappear();
-
-    await transactionPage.clickOnTransactionsMenuButton();
-    await transactionPage.clickOnCreateNewTransactionButton();
-    await transactionPage.clickOnTransferTokensTransaction();
-    await transactionPage.fillInTransferFromAccountId();
-    await transactionPage.fillInTransferAmountFromAccount('10');
-    await transactionPage.fillInTransferToAccountId(receiverAccount);
-    await transactionPage.clickOnAddTransferFromButton();
-    await transactionPage.fillInTransferAmountToAccount('200');
-    await transactionPage.clickOnAddTransferToButton();
-
-    const isButtonEnabled = await transactionPage.isSignAndSubmitButtonEnabled();
-    expect(isButtonEnabled).toBe(false);
-  });
-
-  test('Verify user can execute approve allowance tx', async () => {
-    await transactionPage.ensureAccountExists();
-    const accountFromList = await transactionPage.getFirstAccountFromList();
-    const amountToBeApproved = '10';
-    const transactionId = await transactionPage.approveAllowance(
-      accountFromList,
-      amountToBeApproved,
-    );
-
-    const transactionDetails = await transactionPage.mirrorGetTransactionResponse(
-      transactionId ?? '',
-    );
-    const transactionType = transactionDetails?.name;
-    const result = transactionDetails?.result;
-    expect(transactionType).toBe('CRYPTOAPPROVEALLOWANCE');
-    expect(result).toBe('SUCCESS');
-
-    const isTxExistingInDb = await transactionPage.verifyTransactionExists(
-      transactionId ?? '',
-      'Account Allowance Approve Transaction',
-    );
-
-    expect(isTxExistingInDb).toBe(true);
-  });
-
-  test('Verify all elements are present on file create tx page', async () => {
-    await transactionPage.clickOnTransactionsMenuButton();
-    await transactionPage.clickOnCreateNewTransactionButton();
-    await transactionPage.clickOnFileServiceLink();
-    await transactionPage.clickOnFileCreateTransaction();
-
-    const isAllElementsVisible = await transactionPage.verifyFileCreateTransactionElements();
-    expect(isAllElementsVisible).toBe(true);
-  });
-
-  test('Verify user can execute file create tx', async () => {
-    const { transactionId } = await transactionPage.createFile('test');
-
-    const transactionDetails = await transactionPage.mirrorGetTransactionResponse(
-      transactionId ?? '',
-    );
-    const transactionType = transactionDetails?.name;
-    const result = transactionDetails?.result;
-    expect(transactionType).toBe('FILECREATE');
-    expect(result).toBe('SUCCESS');
-  });
-
-  test('Verify file is stored in the db after file create tx', async () => {
-    await transactionPage.ensureFileExists('test');
-    const fileId = await transactionPage.getFirsFileIdFromCache();
-
-    const isExistingInDb = await transactionPage.verifyFileExists(fileId?? '');
-
-    expect(isExistingInDb).toBe(true);
-  });
-
-  test('Verify user can execute file read tx', async () => {
-    await transactionPage.ensureFileExists('test');
-    const fileId = await transactionPage.getFirsFileIdFromCache();
-    const textFromCache = await transactionPage.getTextFromCache(fileId?? '');
-
-    const readContent = await transactionPage.readFile(fileId?? '');
-
-    expect(readContent).toBe(textFromCache);
-  });
-
-  test('Verify user can execute file update tx', async () => {
-    const newText = 'Lorem Ipsum';
-    await transactionPage.ensureFileExists('test');
-    const fileId = await transactionPage.getFirsFileIdFromCache();
-    const transactionId = await transactionPage.updateFile(fileId?? '', newText);
-
-    const transactionDetails = await transactionPage.mirrorGetTransactionResponse(transactionId?? '');
-    const transactionType = transactionDetails?.name;
-    const result = transactionDetails?.result;
-    expect(transactionType).toBe('FILEUPDATE');
-    expect(result).toBe('SUCCESS');
-
-    //Verify file content is updated
-    const readContent = await transactionPage.readFile(fileId?? '');
-    const textFromCache = await transactionPage.getTextFromCache(fileId?? '');
-    expect(readContent).toBe(textFromCache);
-  });
-
-  test('Verify user can execute file append tx', async () => {
-    const newText = ' extra text to append';
-    await transactionPage.ensureFileExists('test');
-    const fileId = await transactionPage.getFirsFileIdFromCache();
-    const transactionId = await transactionPage.appendFile(fileId?? '', newText);
-
-    const transactionDetails = await transactionPage.mirrorGetTransactionResponse(transactionId?? '');
-    const transactionType = transactionDetails?.name;
-    const result = transactionDetails?.result;
-    expect(transactionType).toBe('FILEAPPEND');
-    expect(result).toBe('SUCCESS');
-
-    //Verify file content is appended
-    const readContent = await transactionPage.readFile(fileId?? '');
-    const textFromCache = await transactionPage.getTextFromCache(fileId?? '');
-    expect(readContent).toBe(textFromCache);
-  });
 });
