+++ automation/tests/local-transactions/workflowHistoryDetailsTests.test.ts
@@ -18,7 +18,7 @@ let transactionPage: TransactionPage;
 let detailsPage: DetailsPage;
 let isolationContext: ActivatedTestIsolationContext | null = null;
 
-test.describe('Workflow history/detail tests @local-transactions', () => {
+test.describe('Workflow history/detail account and transfer tests @local-transactions', () => {
   test.beforeAll(async () => {
     ({ app, window, isolationContext } = await setupLocalSuiteApp(test.info()));
   });
@@ -235,107 +235,4 @@ test.describe('Workflow history/detail tests @local-transactions', () => {
     expect(allowanceAmount).toContain(amountToBeApproved + ' ℏ');
   });
 
-  test('Verify file create tx is displayed in history page', async () => {
-    const { transactionId } = await transactionPage.createFile('test');
-    await transactionPage.clickOnTransactionsMenuButton();
-    await detailsPage.assertTransactionDisplayed(transactionId ?? '', 'File Create');
-  });
-
-  test('Verify transaction details are displayed for file create tx ', async () => {
-    const { transactionId } = await transactionPage.createFile('test');
-    await transactionPage.clickOnTransactionsMenuButton();
-    await detailsPage.clickOnFirstTransactionDetailsButton();
-    await detailsPage.assertTransactionDetails(
-      transactionId ?? '',
-      'File Create',
-    );
-    const isKeyButtonVisible = await detailsPage.isSeeKeyDetailsButtonVisible();
-    expect(isKeyButtonVisible).toBe(true);
-
-    const fileDetailsExpirationTime = await detailsPage.getFileDetailsExpirationTime();
-    expect(fileDetailsExpirationTime).toBeTruthy();
-
-    const isViewContentButtonVisible = await detailsPage.isViewContentsButtonVisible();
-    expect(isViewContentButtonVisible).toBe(true);
-  });
-
-  test('Verify file update tx is displayed in history page', async () => {
-    const newText = 'Lorem Ipsum';
-    await transactionPage.ensureFileExists('test');
-    const fileId = await transactionPage.getFirsFileIdFromCache();
-    const transactionId = await transactionPage.updateFile(fileId ?? '', newText);
-    await transactionPage.clickOnTransactionsMenuButton();
-    await detailsPage.assertTransactionDisplayed(transactionId ?? '', 'File Update');
-  });
-
-  test('Verify transaction details are displayed for file update tx ', async () => {
-    const newText = 'New text';
-    await transactionPage.ensureFileExists('test');
-    const fileId = await transactionPage.getFirsFileIdFromCache();
-    const transactionId = await transactionPage.updateFile(fileId ?? '', newText);
-    await transactionPage.clickOnTransactionsMenuButton();
-    await detailsPage.clickOnFirstTransactionDetailsButton();
-    await detailsPage.assertTransactionDetails(
-      transactionId ?? '',
-      'File Update',
-    );
-    const fileIdFromDetailsPage = await detailsPage.getFileDetailsFileId();
-    expect(fileId).toBe(fileIdFromDetailsPage);
-
-    const isViewContentButtonVisible = await detailsPage.isViewContentsButtonVisible();
-    expect(isViewContentButtonVisible).toBe(true);
-  });
-
-  test('Verify file append tx is displayed in history page', async () => {
-    const newText = ' extra text to append';
-    await transactionPage.ensureFileExists('test');
-    const fileId = await transactionPage.getFirsFileIdFromCache();
-    const transactionId = await transactionPage.appendFile(fileId ?? '', newText);
-    await transactionPage.clickOnTransactionsMenuButton();
-    await detailsPage.assertTransactionDisplayed(transactionId ?? '', 'File Append');
-  });
-
-  // This test is failing in CI environment due to bug in the SDK
-  test('Verify transaction details are displayed for file append tx ', async () => {
-    const newText = ' extra text to append';
-    await transactionPage.ensureFileExists('test');
-    const fileId = await transactionPage.getFirsFileIdFromCache();
-    const transactionId = await transactionPage.appendFile(fileId ?? '', newText);
-    await detailsPage.clickOnFirstTransactionDetailsButton();
-    await detailsPage.assertTransactionDetails(
-      transactionId ?? '',
-      'File Append',
-    );
-    const fileIdFromDetailsPage = await detailsPage.getFileDetailsFileId();
-    expect(fileId).toBe(fileIdFromDetailsPage);
-
-    const isViewContentButtonVisible = await detailsPage.isViewContentsButtonVisible();
-    expect(isViewContentButtonVisible).toBe(true);
-  });
-
-  test('Verify breadcrumb is displayed for transaction group item', async () => {
-    const txDescription = 'test account create tx description';
-    const { newTransactionId } = await transactionPage.createNewAccount({
-      description: txDescription,
-    });
-    await transactionPage.clickOnTransactionsMenuButton();
-    await detailsPage.assertTransactionDisplayed(
-      newTransactionId ?? '',
-      'Account Create',
-      txDescription,
-    );
-
-    await detailsPage.clickOnFirstTransactionDetailsButton();
-
-    const nbItems = await detailsPage.countElements('breadcrumb-item-');
-    expect(nbItems).toBe(2);
-    const item1 = await detailsPage.getBreadCrumbItem(0);
-    const item2 = await detailsPage.getBreadCrumbItem(1);
-    expect(await item1.innerText()).toBe('History');
-    expect(await item2.innerText()).toBe('Account Create Transaction');
-
-    await item1.click();
-    const url = window.url();
-    expect(url).toContain('transactions?tab=History');
-  });
 });
