+++ automation/tests/organization-advanced/organizationTransactionLifecycleTests.test.ts
@@ -2,92 +2,35 @@ import { expect, Page, test } from '@playwright/test';
 import { OrganizationPage, UserDetails } from '../../pages/OrganizationPage.js';
 import { LoginPage } from '../../pages/LoginPage.js';
 import { TransactionPage } from '../../pages/TransactionPage.js';
-import { flushRateLimiter } from '../../utils/db/databaseUtil.js';
-import { setDialogMockState } from '../../utils/runtime/dialogMocks.js';
-import type { TransactionToolApp } from '../../utils/runtime/appSession.js';
 import { waitForValidStart } from '../../utils/runtime/timing.js';
-import { setupOrganizationAdvancedFixture } from '../helpers/fixtures/organizationAdvancedFixture.js';
-import {
-  setupOrganizationSuiteApp,
-  teardownOrganizationSuiteApp,
-} from '../helpers/bootstrap/organizationSuiteBootstrap.js';
-import type { ActivatedTestIsolationContext } from '../../utils/setup/sharedTestEnvironment.js';
 import { createSequentialOrganizationNicknameResolver } from '../helpers/support/organizationNamingSupport.js';
+import { registerOrganizationAdvancedSuiteHooks } from '../helpers/bootstrap/organizationAdvancedSuiteHooks.js';
 
-let app: TransactionToolApp;
 let window: Page;
 let globalCredentials = { email: '', password: '' };
 
 let transactionPage: TransactionPage;
 let organizationPage: OrganizationPage;
 let loginPage: LoginPage;
-let isolationContext: ActivatedTestIsolationContext | null = null;
-let organizationNickname = 'Test Organization';
 
 let firstUser: UserDetails;
 let complexKeyAccountId: string;
 const resolveOrganizationNickname = createSequentialOrganizationNicknameResolver();
 
 test.describe('Organization Transaction status/signing lifecycle tests @organization-advanced', () => {
-  test.slow();
-
-  test.beforeAll(async () => {
-    ({
-      app,
-      window,
-      loginPage,
-      transactionPage,
-      organizationPage,
-      isolationContext,
-    } = await setupOrganizationSuiteApp(test.info()));
-    window.on('console', msg => {
-      const text = msg.text();
-      if (
-        text.includes('[TXD-DBG]') ||
-        text.includes('[SIG-AUDIT-DBG]') ||
-        text.includes('[ORG-USER-DBG]')
-      ) {
-        console.log('[BROWSER]', text);
-      }
-    });
-  });
-
-  test.beforeEach(async ({}, testInfo) => {
-    await flushRateLimiter();
-    await setDialogMockState(window, { savePath: null, openPaths: [] });
-
-    organizationNickname = resolveOrganizationNickname(testInfo.title);
-    const fixture = await setupOrganizationAdvancedFixture(
-      window,
-      loginPage,
-      organizationPage,
-      organizationNickname,
-    );
-    globalCredentials.email = fixture.localCredentials.email;
-    globalCredentials.password = fixture.localCredentials.password;
-    firstUser = fixture.firstUser;
-    complexKeyAccountId = fixture.complexKeyAccountId;
-    await transactionPage.clickOnTransactionsMenuButton();
-
-    await organizationPage.waitForElementToDisappear('.v-toast__text');
-    await organizationPage.closeDraftModal();
-
-    if (process.env.CI) {
-      await new Promise(resolve => setTimeout(resolve, 1000));
-    }
-  });
-
-  test.afterEach(async () => {
-    try {
-      await organizationPage.logoutFromOrganization();
-    } catch {
-      // Several tests delete or fully consume the current org session.
-      // The next beforeEach recreates the fixture from scratch.
-    }
-  });
-
-  test.afterAll(async () => {
-    await teardownOrganizationSuiteApp(app, isolationContext);
+  registerOrganizationAdvancedSuiteHooks({
+    resolveOrganizationNickname,
+    onSuiteReady: suite => {
+      ({ window, loginPage, transactionPage, organizationPage } = suite);
+    },
+    getPages: () => ({ window, loginPage, transactionPage, organizationPage }),
+    onFixtureReady: fixture => {
+      globalCredentials.email = fixture.localCredentials.email;
+      globalCredentials.password = fixture.localCredentials.password;
+      firstUser = fixture.firstUser;
+      complexKeyAccountId = fixture.complexKeyAccountId;
+    },
+    logoutFromOrganization: () => organizationPage.logoutFromOrganization(),
   });
 
   test('Verify transaction is shown "Ready for Execution" and correct stage is displayed', async () => {
